Seeing as I am tired I am stopping for the day but I accidently put problem three in d.py so I moved it to c.py
